refactor(TransactionCard): Separate hooks and styles
