+++ src/hooks/useTxForAddress.ts
@@ -8,11 +8,6 @@ const useTxForAddress = ({ address, chain }: { address: string; chain: string })
   const [error, setError] = React.useState<any>(undefined);
   const [loading, setLoading] = React.useState(true);
 
-  React.useEffect(() => {
-    if (!address) return;
-    fetchTx();
-  }, [address]);
-
   const fetchTx = React.useCallback(async () => {
     setLoading(true);
     setError(undefined);
@@ -36,6 +31,11 @@ const useTxForAddress = ({ address, chain }: { address: string; chain: string })
     } finally {
       setLoading(false);
     }
+  }, [address]);
+
+  React.useEffect(() => {
+    if (!address) return;
+    fetchTx();
   }, [address]);
 
   return {
