+++ src/pages/Account/components/TransactionsCard/components/TabButton/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Fab from '@mui/material/Fab';
 import { Typography } from '@mui/material';
+import useStyles from './useStyles';
 
 type Props = {
   /**
@@ -23,24 +24,11 @@ type Props = {
  * A tab button, for use in the TransactionCard component
  */
 const TabButton = ({ label, onClick, isSelected }: Props) => {
+  const { tabButton } = useStyles(isSelected);
+
   return (
-    <Fab
-      sx={{
-        width: '30%',
-        margin: 0.5,
-        color: isSelected ? 'text.primary' : 'text.secondary',
-        backgroundColor: isSelected ? 'primary.main' : 'grey.200',
-        '&:hover': {
-          backgroundColor: isSelected ? 'primary.main' : 'grey.200',
-          opacity: 0.8,
-        },
-      }}
-      size="small"
-      variant="extended"
-      arial-label={label}
-      onClick={onClick}
-    >
-      <Typography sx={{ fontSize: 'small' }}>{label}</Typography>
+    <Fab sx={tabButton} size="small" variant="extended" arial-label={label} onClick={onClick}>
+      <Typography fontSize="small">{label}</Typography>
     </Fab>
   );
 };
